+++ db/migrations/knex/20200324123515_create_anonymous_role.js
@@ -1,15 +1,13 @@
-exports.up = function (knex) {
-  return knex.raw(`
+exports.up = knex =>
+  knex.raw(`
     CREATE ROLE anonymous NOINHERIT NOLOGIN;
 
     GRANT USAGE ON SCHEMA api, auth, public TO anonymous;
-  `);
-};
+  `)
 
-exports.down = function (knex) {
-  return knex.raw(`
+exports.down = knex =>
+  knex.raw(`
     REVOKE USAGE ON SCHEMA api, auth, public FROM anonymous;
 
     DROP ROLE anonymous;
-  `);
-};
+  `)