+++ db/seeds/dev/03_books.js
@@ -1,36 +1,36 @@
 exports.seed = async knex => {
-  await knex("api.books").insert([
+  await knex('api.books').insert([
     {
-      title: "An Analysis of Betty Friedan's The Feminine Mystique",
       author_id: 5,
+      title: "An Analysis of Betty Friedan's The Feminine Mystique",
     },
     {
-      title: "Anna Karenina",
       author_id: 4,
+      title: 'Anna Karenina',
     },
     {
-      title: "Crow Blue",
       author_id: 3,
+      title: 'Crow Blue',
     },
     {
-      title: "Immortality",
       author_id: 2,
+      title: 'Immortality',
     },
     {
-      title: "Symphony in White",
       author_id: 3,
+      title: 'Symphony in White',
     },
     {
-      title: "The Prophet",
       author_id: 1,
+      title: 'The Prophet',
     },
     {
-      title: "The Unbearable Lightness of Being",
       author_id: 2,
+      title: 'The Unbearable Lightness of Being',
     },
     {
-      title: "War and Peace",
       author_id: 4,
+      title: 'War and Peace',
     },
-  ]);
-};
+  ])
+}