+++ src/index.d.ts
@@ -1,5 +1,0 @@
-import { PostgresterStatic } from "./types";
-
-declare const postgrester: PostgresterStatic;
-
-export = postgrester;