+++ src/texts.ts
@@ -1,4 +1,4 @@
 export default {
   POSTGRESTER_DEPRECATION_BASE_URI:
-    "Warning: Using `baseUri` is now deprecated and will removed in v2.0.0: https://github.com/ivangabriele/postgrester/wiki/Deprecations#baseuri-option",
-};
+    'Warning: Using `baseUri` is now deprecated and will removed in v2.0.0: https://github.com/ivangabriele/postgrester/wiki/Deprecations#baseuri-option',
+}